Separated TestConfiguration class from ConfigScreen.dart
